% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{addtotal.list}
\alias{addtotal.list}
\title{Add total}
\usage{
addtotal.list(x)
}
\arguments{
\item{x}{the zoo objects}
}
\description{
Calculate the total of a list of zoo objects and as a new zoo object called total
}
\examples{
addtotal.list(list(zoo(c(1,2),c(3,4)),zoo(c(2,3),c(4,5))))
}
\keyword{convenience}
