% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{firstinvec}
\alias{firstinvec}
\title{first in vector}
\usage{
firstinvec(x)
}
\arguments{
\item{x}{is the vector}
}
\description{
convenience function to return first element in a vector
}
\examples{
firstinvec(1:3)
}
\keyword{convenience}
