% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{formatwblank}
\alias{formatwblank}
\title{format a vector of numbers}
\usage{
formatwblank(x, digits = 0, addtxt = NULL, big.mark = ",")
}
\arguments{
\item{x}{is the vector of numbers}

\item{digits}{is the number of digits for rounding}

\item{addtxt}{is to text to add after each number}

\item{big.mark}{defaults to ","}
}
\description{
uses blankfun with sapply
}
\examples{
formatwblank(sqrt(1:4),2)
}
\keyword{convenience}
