% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv_functionsss.R
\name{fv.tr}
\alias{fv.tr}
\title{FV with time varying interest rates}
\usage{
fv.tr(CF, TR, ValDate = NULL)
}
\arguments{
\item{CF}{is the cash flow zoo object}

\item{TR}{is the total return time series}

\item{ValDate}{is the valuation date defaults to last date in time series}
}
\description{
calculates FV from cash flow and a total return index
}
\keyword{present_value}
