% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv_functionsss.R
\name{fv.z}
\alias{fv.z}
\title{Future Value Function}
\usage{
fv.z(i, cf, freq = 1, apr = FALSE, now = NULL,
  drop.after.now = TRUE, adjdisc = NULL)
}
\arguments{
\item{i}{is the periodic interest rate}

\item{cf}{is the cash flow, a zoo object or vector}

\item{apr}{if TRUE means the periodic/annual interest rate will be adjusted to the payment periodicity by division, otherwise geometric}

\item{now}{is the date for valuation}

\item{drop.after.now}{if TRUE drops cash flows after now}

\item{adjdisc}{is there for a forgotten reason}
}
\description{
Calculates future value at end of cash flow or date "now"
}
\examples{
fv.z(.1,zoo(1:10,11:20))
}
