% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{gd}
\alias{gd}
\title{Growth of a dollar}
\usage{
gd(x)
}
\arguments{
\item{x}{an arithmetic return series}
}
\description{
Growth of a dollar of an arithmetic return series
}
\examples{
gd(c(.1,.2))
}
\keyword{convenience}
