% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{gdweight}
\alias{gdweight}
\title{Growth of a dollar weight}
\usage{
gdweight(x)
}
\arguments{
\item{x}{an arithmetic return series}
}
\description{
Growth of a dollar of an arithmetic return series returns on ly the return part
}
\examples{
gdminus1(c(.1,.2))
}
\keyword{convenience}
