% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_functions.R
\name{gpcomp}
\alias{gpcomp}
\title{general partner compensation}
\usage{
gpcomp(dmat, ret, capital = 100, invcost = 100)
}
\arguments{
\item{dmat}{a data frame describing the deal structure with columns for asset management fee, pref, catchup and carry}

\item{ret}{one or more returns as a vector}

\item{capital}{amount contributed}

\item{invcost}{cost of investments (may be different from cost because of asset management fee)}
}
\description{
a function to evaluate gp compensation for one or more returns
}
\examples{
peinv=100
pecap=102
dmat.pe=data.frame(am=0,pref=(.08*pecap),catchup=.5,carry=.2)
ans.pe=gpcomp(dmat.pe,ret=seq(100,130,.1),invcost=peinv,capital=pecap)
}
\keyword{private_equity}
