% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv_functionsss.R
\name{grow}
\alias{grow}
\title{constant growth}
\usage{
grow(c = 1, t = term, g)
}
\arguments{
\item{c}{is starting value}

\item{t}{is number of periods to grow}

\item{g}{is the growth rate}
}
\description{
create a vector with constant rate of change
}
\examples{
grow(1,10,.05)
}
\keyword{present_value}
