% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv_functionsss.R
\name{grow.z}
\alias{grow.z}
\title{constant growth}
\usage{
grow.z(r, c, n, freq = 1, start = 1, retclass = "zoo")
}
\arguments{
\item{r}{is the rate of growth}

\item{c}{is the starting value}

\item{n}{is the number of periods}

\item{freq}{is the frequency 1 for annual, 12 for monthly, etc}

\item{start}{is the starting date}

\item{retclass}{defaults to zoo}
}
\description{
constant growth with option to return a zooreg object
}
\examples{
grow.z(.1,1,10)
}
\keyword{present_value}
