% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_functions.R
\name{incentfee.rev}
\alias{incentfee.rev}
\title{incentive fee reverse engineer}
\usage{
incentfee.rev(nav, cf, cu, pref = 0.08, cry = 0.2, fre = 4)
}
\arguments{
\item{nav}{is a time series of navs}

\item{cf}{is a time series of cash flows}

\item{cu}{is the catchup parameter}

\item{pref}{is the preferred return defaults to 8}

\item{cry}{is the carry defaults to 20}

\item{fre}{is the compounding period defaults to quarterly}
}
\description{
reverse engineer incentive fee from known performance data
}
\examples{
cf=zoo(c(-100,-100),as.Date("2018-1-1")+c(1,90))
nav=zoo(c(300)as.Date("2018-6-30"))
incentfee.rev=function(nav,cf,1)
}
\keyword{private_equity}
