% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr_functions.R
\name{irr.z}
\alias{irr.z}
\title{IRR}
\usage{
irr.z(cf.z, gips = FALSE)
}
\arguments{
\item{cf}{is the cash flow, a zoo object with date (i.e. daily) class time}
}
\description{
Calculate an IRR from a zoo object with GIPS compliance option
}
\examples{
irr.z(zoo(c(-1,1.05),as.Date("2018-1-1")+c(0,365)))
}
\keyword{IRR}
