% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr_functions.R
\name{irrgap.z}
\alias{irrgap.z}
\title{IRR gap}
\usage{
irrgap.z(i, cf.z, end = NA)
}
\arguments{
\item{i}{is the target IRR}

\item{cf}{is the cash flow, a zoo object with date (i.e. daily) class time}
}
\description{
How much more (or less) terminal value to achieve target IRR
}
\examples{
irrgap.z(.1,zoo(c(-1,1.05),as.Date("2018-1-1")+c(0,365)))
}
\keyword{IRR}
