% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{lastinvec}
\alias{lastinvec}
\title{last elecment in vector}
\usage{
lastinvec(x, naremove = FALSE)
}
\arguments{
\item{x}{is the vector}

\item{naremove}{changes trailing NA to zero if set to true}
}
\description{
convenience function to grab the last element in a vector
}
\examples{
lastinvec(1:3)
}
\keyword{convenience}
