% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loan_functions.R
\name{loanamort}
\alias{loanamort}
\title{Loan amortization function}
\usage{
loanamort(r = NULL, bal0 = NULL, pmt = NULL, n = NULL,
  apr = FALSE, start = NULL, freq = 1)
}
\arguments{
\item{r}{is the interest rate}

\item{bal0}{is the balance at time 0 (can be a zooreg object)}

\item{pmt}{is the periodic payment}

\item{n}{is the nuber of periods}

\item{apr}{if TRUE means the conversion of rate in r to frequency is done by arithmetic division, otherwise geometric}

\item{start}{is optional starting date}

\item{freq}{is the frequency of payments - 1 means the frequency is the same as the stated interest rate}
}
\description{
Returns value for the missing parameter and full amortization schedule for the loan as zooreg objects
}
\examples{
loanamort(r=.1, bal0=1000, pmt=NULL, n=30, start=2018, freq=12)
}
\keyword{loan}
