% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loan_functions.R
\name{lpmt}
\alias{lpmt}
\title{Loan payment}
\usage{
lpmt(bal, i, n)
}
\arguments{
\item{bal}{is initial balance}

\item{i}{is interest rate}

\item{n}{is number of periods}
}
\description{
Returns level payment amount given initial balance, interest rate and number of payments
}
\examples{
lpmt(1000,.1,30)
}
\keyword{loan}
