% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{make.table}
\alias{make.table}
\title{zoo list to a table}
\usage{
make.table(x, by = "year", time.horizontal = TRUE, fun = sum)
}
\arguments{
\item{x}{is the zoo list}

\item{by}{is "year", "quarter" or "month"}

\item{time.horizontal}{makes time as the columns if TRUE otherwise as rows}

\item{fun}{is function for aggregating by time frame, defaults to sum}
}
\description{
make a table out of a list of zoo objects organized by year, month or quarter
}
\examples{
cf1=zoo(1:20, as.Date("2018-1-1")+(1:20*40))
cf2=zoo(1:20, as.Date("2018-1-1")+(1:20*50))
x=list(cf1,cf2)
make.table(x)
}
\keyword{convenience}
