% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{merge0}
\alias{merge0}
\title{Merge zero}
\usage{
merge0(...)
}
\description{
Merge with a fill=0 parameter
}
\examples{
merge0(list(zoo(c(1,2),c(3,4)),zoo(c(2,3,4),c(4,5,6))))
}
\keyword{convenience}
