% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{mergelist.z}
\alias{mergelist.z}
\title{Merge list}
\usage{
mergelist.z(z.list)
}
\arguments{
\item{x}{the zoo objects}
}
\description{
Merge a list of zoo objects into a zoo data.frame
}
\examples{
mergelist.z(list(zoo(c(1,2),c(3,4)),zoo(c(2,3),c(4,5))))
}
\keyword{convenience}
