% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr_functions.R
\name{npv.znoadjust}
\alias{npv.znoadjust}
\title{NPV for irr.z function}
\usage{
npv.znoadjust(i, cf.z, freq, tdiff)
}
\arguments{
\item{i}{is the annual interest rate}

\item{cf.z}{is a zoo object cash flow}

\item{freq}{is the compounding period (e.g. 365 for daily, 12 for monthly)}

\item{tdiff}{is the previously calculated time difference across cf.z}
}
\description{
a concise NPV function used by irr.z (everything extra stripped out since function called a bazillion times)
}
