% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldstyle_performance.R
\name{payback}
\alias{payback}
\title{payback}
\usage{
payback(cf)
}
\arguments{
\item{cf}{is the cash flow with even periodicity}
}
\description{
payback period (how long before you get your money back)
}
\examples{
tvpi(c(-1,1.1))
}
\keyword{Performance}
