% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_functions.R
\name{pe.performance.return.index}
\alias{pe.performance.return.index}
\title{pe performance with cash flow data from file}
\usage{
pe.performance.return.index(cf.filename, bench.ticker.vec = c("^RUT", "SPY"),
  useallvals = FALSE, service = "yahoo")
}
\arguments{
\item{cf.filename}{the name of the file}

\item{bench.ticker.vec}{a vector of tickers}

\item{useallvals}{limts to "C" and "V" factors if FALSE}

\item{service}{is "yahoo" or "bloomberg"}
}
\description{
pe performance with cash flow data from file
}
\examples{
file="filename"
pe.performance.return.index(file)
}
\keyword{private_equity}
