% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_functions.R
\name{pe.performance.roll}
\alias{pe.performance.roll}
\title{evolution of pe performance}
\usage{
pe.performance.roll(cf, val, idx)
}
\arguments{
\item{cf}{is a cash flow zoo object}

\item{val}{is zoo object of valuations}

\item{idx}{ia a zoo object total return index}
}
\description{
evolution of pe performance
}
\examples{
dates=as.Date("2018-1-1")+c(1,31,61)
cf=zoo(c(-100,5,10),dates)
ind=zoo(c(100,101,102),dates)
val=zoo(c(100,100),dates[-1])
pestats(cf,val,ind)
}
\keyword{private_equity}
