% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_functions.R
\name{pestats}
\alias{pestats}
\title{pe performance}
\usage{
pestats(cf, ind)
}
\arguments{
\item{cf}{is cashflow zoo object}

\item{ind}{is a total return index zoo object}
}
\description{
calculate pe performance metrics
}
\examples{
dates=as.Date("2018-1-1")+c(1,31,61)
cf=zoo(c(-100,5,110),dates)
ind=zoo(c(100,101,102),dates)
pestats(cf,ind)
}
\keyword{private_equity}
