% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_functions.R
\name{pfd.return}
\alias{pfd.return}
\title{Preferred return}
\usage{
pfd.return(cf, int, freq = 4, mdate = NA)
}
\arguments{
\item{cf}{is a time series cash flow}

\item{int}{is the preferred return rate}

\item{freq}{is the compouding period defaults to quarterly}

\item{mdate}{is the measurement date}
}
\description{
calculates preferred return from cash flow
}
\examples{
cf=zoo(c(-100,-100),as.Date("2018-1-1")+c(1,90))
pfd.return(cf,.08)
}
\keyword{private_equity}
