% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loan_functions.R
\name{prinportion}
\alias{prinportion}
\title{Principal portion}
\usage{
prinportion(bal, i, pmt, n)
}
\arguments{
\item{bal}{is the initial balance}

\item{i}{is the interest rate}

\item{pmt}{is the period payment}

\item{n}{is the period for evaluation(can be a vector)}
}
\description{
Returns the principal portion of the nth payment
}
\examples{
intportion(1000,.1,106.08,c(9,10))
}
\keyword{loan}
