% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loan_functions.R
\name{simple.interest.z}
\alias{simple.interest.z}
\title{Simple interest}
\usage{
simple.interest.z(i, cf, start = time(cf)[1], end = lastinvec(time(cf)))
}
\arguments{
\item{i}{is the annual interest rate (an apr, will be divided by 365 to get a daily rate)}

\item{cf}{is the cash flow, a zoo object}

\item{start}{is the start day for the calculation (default is first day in cf)}

\item{end}{is the end day for the calculation (default is last day in cf)}
}
\description{
A function to calculate simple interest on a loan (handy for preferred return calculations with annual or quarterly compounding terms)
}
\examples{
simple.interest.z(.1, zoo(c(1,2),as.Date("2018-1-1")+c(3,5)),start=as.Date("2018-1-1"),end=as.Date("2018-1-31"))

}
\keyword{loan}
