% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_functions.R
\name{testans}
\alias{testans}
\title{test gpcomp}
\usage{
testans(ans)
}
\arguments{
\item{ans}{the result of a call to gpcomp}
}
\description{
a convenience function to test the result of gpcomp
}
\examples{
testans(ans)
}
\keyword{private_equity}
