% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldstyle_performance.R
\name{tvpi}
\alias{tvpi}
\title{tvpi}
\usage{
tvpi(cf)
}
\arguments{
\item{cf}{is the cash flow}
}
\description{
calcultes TVPI (total value as percent of paid in)
}
\examples{
tvpi(c(-1,1.1))
}
\keyword{Performance}
