% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{unzoo}
\alias{unzoo}
\title{unzoo}
\usage{
unzoo(
  zoo.object,
  wide = TRUE,
  dataname = "Value",
  datename = "Date",
  category = "Categories"
)
}
\arguments{
\item{zoo.object}{is the zoo object}

\item{wide}{starting form is wide if true}

\item{dataname}{"Value" is default}

\item{datename}{"Date" is default}

\item{category}{"Categories is default}
}
\description{
convert a zoo data.frame to regular data frame with date as a column
}
\examples{
unzoo(merge0(list(zoo(c(1,2),c(3,4)),zoo(c(2,3,4),c(4,5,6))))) 

}
\keyword{convenience}
