% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_functions.R
\name{wf}
\alias{wf}
\title{waterfall}
\usage{
wf(w, c)
}
\arguments{
\item{w}{the waterfall as vector}

\item{c}{the amount of cash to distribute}
}
\description{
given waterfall w in dollars and available cash c, distribute the cash to the waterfall
}
\examples{
wf(c(1,2,3),5)
}
\keyword{private_equity}
