% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{window.list}
\alias{window.list}
\title{window a list}
\usage{
\method{window}{list}(x, start, end)
}
\arguments{
\item{x}{is the zoo list}

\item{start}{is the start date}

\item{end}{is the end date}
}
\description{
apply window to all members of a list of zoo objects
}
\examples{
cf1=zoo(1:20, as.Date("2018-1-1")+(1:20*40))
cf2=zoo(1:20, as.Date("2018-1-1")+(1:20*50))
x=list(cf1,cf2)
window.list(x,as.Date("2018-1-1),as.Date("2018-12-31))
}
\keyword{convenience}
